Find sections in the seating tool that are no longer linked to Brightspace rosters
This will happen if a roster is removed from the Brightspace side.

Right now the delete is turned off, and this case will result in a
RuntimeException being thrown during the sync.  Once we're confident
that we're only deleting the stuff we want to delete, we'll turn the
option on and start cleaning up detached rosters.