Add write-ahead logging for changed section tracking
We rely on tracking sections with changed membership in order to
resync the appropriate seating tool sites after the job finishes.
However, if the CmSynchronizer job is interrupted during its run (due
to a server restart, for example), the sections updated so far are not
recorded.

To get around this, use a WAL to store updated sections durably.